Fix broken SRAM saving in latest build
Turns out merging destType and sourceType into a single variable the way I did was a mistake as it got overwritten in a subroutine of the IncrementSectorNum macro, or even as early as in the CardWriteFile routine. While it would indeed be possible to use just one variable (i.e., 1 byte), the additional bitwise operations needed in every iteration of NextCluster would probably slow down game loading quite a bit, so this idea has been discarded for now.

Also, remove yet another completely unused variable (bankOffset)