Remove some redundant ROM size checks, and make ROM mapping detection code more readable
Also, this includes a first attempt at getting ExLoROM games to work on the PowerPak. As it turns out, it used to fetch a wrong (HiROM) entry from the banking database when trying to load e.g. a 48 Mbit ExLoROM title. ExLoROM games still don't work, but hey, it's a start. ;-)

Also, remove the useless HiROM >32 Mbit entries from the database